ユーザー情報をログインしたときにシェアードプリファレンスに保存
